lowering: reorganize module hierachy + module-level docs
