+++ lowering/src/linear_scan.rs
@@ -1,9 +1,11 @@
+//! Linear-scan register allocation.
+
 use super::{
     lir, live_variable_analysis,
     register::{Amd64Reg, RegisterAllocator},
     VarId,
 };
-use crate::{amd64, lir_allocator::Ptr};
+use crate::allocator::Ptr;
 use gcollections::ops::bounded::Bounded;
 use interval::Interval;
 use live_variable_analysis::{LVAResult, ScheduledInstr};
@@ -48,7 +50,7 @@ pub(crate) fn register_allocation(func: &mut lir::Function, lva_result: LVAResul
     let var_live = live_ranges_by_start; // fixme
 
     // FIXME stack calling convention dropped
-    let reg_alloc = RegisterAllocator::new(func.nargs, amd64::CallingConv::X86_64);
+    let reg_alloc = RegisterAllocator::new(func.nargs, crate::CallingConv::X86_64);
     let mut linear_scan = LinearScanAllocator::new(reg_alloc, var_live, &lsa_params_list);
     linear_scan.run(&scheduled_instrs);
     linear_scan.into()
